+++ src/adRendering.ts
@@ -19,7 +19,7 @@ import { hook } from './hook.js';
 import { fireNativeTrackers } from './native.js';
 import adapterManager from './adapterManager.js';
 import { useMetrics } from './utils/perfMetrics.js';
-import { filters } from './targeting.js';
+import { filters } from './targeting/filters.js';
 import { EVENT_TYPE_WIN, parseEventTrackers, TRACKER_METHOD_IMG } from './eventTrackers.js';
 import type { Bid } from "./bidfactory.ts";
 import { yieldsIf } from "./utils/yield.ts";
