+++ src/prebid.ts
@@ -28,7 +28,8 @@ import { listenMessagesFromCreative } from './secureCreatives.js';
 import { userSync } from './userSync.js';
 import { config } from './config.js';
 import { auctionManager } from './auctionManager.js';
-import { isBidUsable, targeting } from './targeting.js';
+import { isBidUsable } from './targeting/filters.js';
+import { targeting } from './targeting.js';
 import { hook, wrapHook } from './hook.js';
 import { loadSession } from './debugging.js';
 import { storageCallbacks } from './storageManager.js';
